Added test for comment. It doesn't need e2e test so I move on.
